refactor(python): split server.py into tool modules
server.py (914 lines) split into: server.py entry point (64 lines), tools_prts (6 PRTS wiki tools), tools_gamedata (16 game-data tools), tools_story (8 story tools), startup_sync (sync orchestration + _require_story_zip). Sync symbols re-exported from server for backward compat with startup-sync tests.

test_tool_surface updated to walk nested defs (tools now nested in register_* wrappers). All 194 tests pass.
